upload forgot password part1

+++ client/src/pages/Auth/ResetPassword.jsx
@@ -5,9 +5,12 @@ import { useState } from 'react';
 import newRequest from '../../utils/newRequest';
 import { toast } from 'react-toastify';
 import axios from 'axios';
+import useForm from 'react-hook-form';
 
 const ResetPassword = () => {
+    const { register, errors, watch } = useForm({});
     const [password, setPassword] = useState('');
+    const [confirmPassword, setConfirmPassword] = useState('');
     const navigate = useNavigate();
     const { id, token } = useParams();
 
@@ -15,8 +18,12 @@ const ResetPassword = () => {
     const handleSubmit = async (e) => {
         e.preventDefault();
 
+        if (password !== confirmPassword) {
+            return toast('Passwords do not match');
+        }
+
         try {
-            const res = await newRequest.post(`/auth/reset-password/${id}/${token}`, { password });
+            const res = await newRequest.post(`/auth/reset-password/${token}`, { password });
             if (res && res.data.success) {
                 toast.success(res.data && res.data.message);
                 localStorage.setItem('currentUser', JSON.stringify(res.data));
@@ -47,7 +54,28 @@ const ResetPassword = () => {
                             placeholder="name@example.com"
                             value={password}
                             onChange={(e) => setPassword(e.target.value)}
+                            ref={register({
+                                required: 'You must specify a password',
+                                minLength: {
+                                    value: 8,
+                                    message: 'Password must have at least 8 characters',
+                                },
+                            })}
                         />
+                        {errors.password && <p>{errors.password.message}</p>}
+                    </FloatingLabel>
+                    <FloatingLabel controlId="floatingEmail" label="Confirm Password" className="mb-3">
+                        <Form.Control
+                            required
+                            type="password"
+                            placeholder="name@example.com"
+                            value={confirmPassword}
+                            onChange={(e) => setConfirmPassword(e.target.value)}
+                            ref={register({
+                                validate: (value) => value === password || 'The passwords do not match',
+                            })}
+                        />
+                        {errors.password_repeat && <p>{errors.password_repeat.message}</p>}
                     </FloatingLabel>
                     <Button variant="dark" type="submit" className="w-100 mb-3">
                         SEND
